Starship Troopers Ender's Game Connection.
Starship Troopers is the biographical novel and movie of Mazer Rackham. During these wars, military commanders fought with a mentality of "kill anything we see". Blanket fire covered up the infiltration of humans through parasetic infiltration. The parasites are not talked about it the novels or movies because humans were unaware of their existance.

We are now aware of quantum bonds between "brain" and arachnid-like muscularskeleton structure. We can use this to launch targeted, distrubuted attacks to "brain" portions of the species. Scanning technology should be used to identify hosts which have been overtaken by the parasite creature. Hosts infected with the parasite operate from a hive mind which the creature originally spawned from. These Polyps are quantumly connected neural structures that retain control of the species even when physically unatached.