Identified Flying Objects
A competent military would realize the seriousness of the situation and stop telling people this is "unidentified aerial phenomena". These are verified hostile enemy ships. Miliaries look like school children when they cannot even tell the public what they saw. It is shameful to be so unknowledgeable about a defense topic so crucial to national security.