Update README.md
The Pycnogonids fly egg-shaped craft containing human hosts hundreds of thousands of kidnapped humans contained in stasis undiced by blood harvesting. The color of the craft has been identified as "tooth yellow" and has a clear dome on the upper aft portion. The crafts hover above the ocean at underwater stations, and a red tail attaches to some type of deep-sea cable which has been identified as leading to Antarctica.